23. [MainQuest_06] 나만의 AI앱을 플러터로 구현하기(feat.fastAPI)/fastapi_server/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
requests==2.26.0
python-dotenv==0.19.1
openai==0.28.0
